refactor: extract commands module
