+++ src/commands/scan-reminders.ts
@@ -1,0 +1,9 @@
+import type ReminderPlugin from 'main';
+
+export function scanReminders(checking: boolean, plugin: ReminderPlugin): boolean {
+  if (checking) {
+    return true;
+  }
+  plugin.reloadAllFiles();
+  return true;
+}
