+++ src/commands/show-reminder-list.ts
@@ -1,0 +1,8 @@
+import type ReminderPlugin from 'main';
+
+export function showReminderList(checking: boolean, plugin: ReminderPlugin) {
+  if (!checking) {
+    plugin.showReminderList();
+  }
+  return true;
+}
